Revolutionize the way we can actually use ember inspector
Let’s face it, it was a falling out. As the cool summer mornings gave
way to blistering hot summer days, we all felt it approaching. Semantic
and Ember Inspector had a horrendous battle to the death that left
Semantic pushing our application to the bottom of the DOM. But on this
day, on this day indeed, victory is OURS! With great stride we hold on
to control of the DOM and simple move our application to the top again
after every render. Victory to us, my little DOM node friends, victory
to us.