Initial commit for #82.
Added custom-fields route.
Added customFields as a pass-through object in the employee model and
added employeeCustomFields as an array on the company model.
Styled the .no-info-header on the employee page so that it looks good
on mobile when there is no information in the direct contact, emergency
contact, anniversaries, and mailing address cards.